+++ server/src/entities/PropertGroup.ts
@@ -3,7 +3,7 @@ import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 
 import { IProperty } from '@interfaces/Property';
 import { PropertyEntity } from './Property';
-import { IPropertyGroup, PropertyGroupOptionType } from '@interfaces/PropertyGroup';
+import { IPropertyGroup } from '@interfaces/PropertyGroup';
 
 export class ProductGroupBaseEntity implements IPropertyGroup {
     @PrimaryGeneratedColumn('rowid')
@@ -13,10 +13,6 @@ export class ProductGroupBaseEntity implements IPropertyGroup {
     @Column({ unique: true })
     @ApiProperty({ required: false })
     title: string;
-
-    @Column({ select: false })
-    @ApiProperty({ enum: PropertyGroupOptionType, required: false })
-    type: PropertyGroupOptionType;
 
     @Column({ unique: true })
     @ApiProperty({ required: false })
