+++ server/src/interfaces/Filter.ts
@@ -1,12 +1,23 @@
 import { IProperty } from './Property';
 import { IPropertyGroup } from './PropertyGroup';
 
+export interface IFilters {
+    price: any
+    other: Array<IFilterGroup>
+}
+
 export interface IFilterGroup extends IPropertyGroup {
-    properties: Array<IFilter> | null
+    type: FilterType
+    properties: Array<IFilter>
 }
 
 export interface IFilter extends IProperty {
     selected: boolean
     amount: number
     available: number
+}
+
+export enum FilterType {
+    List = 'list',
+    Range = 'range',
 }
