+++ server/src/services/Property.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, Repository } from 'typeorm';
 
-import { CreatePropertyDTO, CreateFilterDTOConstructor } from '@dto/Property';
+import { CreatePropertyDTO, CreatePropertyDTOConstructor } from '@dto/Property';
 import { PropertyEntity } from '@entities/Property';
 import { IProperty } from '@interfaces/Property';
 
@@ -13,7 +13,7 @@ export class PropertyService {
 	) {}
 
 	async createProperty(property: CreatePropertyDTO): Promise<void> {
-		await this.propertyRepo.save(new CreateFilterDTOConstructor(property));
+		await this.propertyRepo.save(new CreatePropertyDTOConstructor(property));
 	}
 
 	getProperty(propertyId: number): Promise<IProperty> {
